Modularise chat function to a separate component
