+++ frontend/src/pages/Collaboration/Collaboration.js
@@ -1,5 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
+import Chat from '../../components/Collaboration/Chat';
 import io from 'socket.io-client';
 
 const Collaboration = () => {
@@ -11,8 +12,6 @@ const Collaboration = () => {
   const hostId = location.state?.hostId;
   const matchedHostId = location.state?.matchedHostId;
 
-  const [message, setMessage] = useState('');
-  const [messages, setMessages] = useState([]);
 
   const socket = io('http://localhost:3002');
 
@@ -20,20 +19,7 @@ const Collaboration = () => {
     // Join the Socket.io room when the component mounts
     socket.emit('joinRoom', { room: roomId, host: hostId });
 
-    // Listen for incoming messages from the server
-
-    socket.on('message', (message) => {
-      console.log('message:', message)
-      setMessages((prevMessages) => [...prevMessages, message]);
-    });
   }, []);
-
-  const sendMessage = () => {
-    if (message) {
-      socket.emit('chatMessage', { room: roomId, message });
-      setMessage('');
-    }
-  };
 
   const handleLeaveRoom = () => {
     socket.emit('leaveRoom', { room: roomId, host: hostId });
@@ -44,23 +30,8 @@ const Collaboration = () => {
     <div>
       <h1>RoomID: {roomId}</h1>
       <h2>You have been match with Host {matchedHostId}</h2>
-      <div>
-        <ul>
-          {messages.map((msg, index) => (
-            <li key={index}>{msg}</li>
-          ))}
-        </ul>
-      </div>
-      <div>
-        <input
-          type="text"
-          placeholder="Enter a message"
-          value={message}
-          onChange={(e) => setMessage(e.target.value)}
-        />
-        <button onClick={sendMessage}>Send</button>
-        <button onClick={handleLeaveRoom}>Leave Room</button>
-      </div>
+      <Chat socket={socket} roomId={roomId} />
+      <button onClick={handleLeaveRoom}>Leave Room</button>
     </div>
   );
 };
